+++ src/config.rs
@@ -1,6 +1,5 @@
 use clap::Parser;
-use ethers::signers::{coins_bip39::English, LocalWallet, MnemonicBuilder, Signer, WalletError};
-use std::{fmt::Debug, fs};
+use std::fmt::Debug;
 
 use crate::Chain;
 
@@ -12,18 +11,6 @@ pub struct TxEnvCLIConfig {
     #[arg(long, env)]
     pub tx_provider_http_endpoint: Option<String>,
 
-    /// Signer mnemonic, overrides `tx_mnemonic_file`
-    #[arg(long, env)]
-    pub tx_mnemonic: Option<String>,
-
-    /// Signer mnemonic file path
-    #[arg(long, env)]
-    pub tx_mnemonic_file: Option<String>,
-
-    /// Mnemonic account index
-    #[arg(long, env)]
-    pub tx_mnemonic_account_index: Option<u32>,
-
     /// Chain ID
     #[arg(long, env)]
     pub tx_chain_id: Option<u64>,
@@ -32,7 +19,7 @@ pub struct TxEnvCLIConfig {
     #[arg(long, env)]
     pub tx_chain_is_legacy: Option<bool>,
 
-    /// Path to tx manager database file
+    /// Path to tx-manager database file
     #[arg(long, env)]
     pub tx_database_path: Option<String>,
 
@@ -49,7 +36,6 @@ pub struct TxEnvCLIConfig {
 pub struct TxManagerConfig {
     pub default_confirmations: usize,
     pub provider_http_endpoint: String,
-    pub wallet: LocalWallet,
     pub chain_id: u64,
     pub chain_is_legacy: bool,
     pub database_path: String,
@@ -63,7 +49,6 @@ impl Debug for TxManagerConfig {
             .field("provider_http_endpoint", &self.provider_http_endpoint)
             .field("chain_id", &self.chain_id)
             .field("chain_is_legacy", &self.chain_is_legacy)
-            .field("wallet_address", &self.wallet.address())
             .field("database_path", &self.database_path)
             .field("gas_oracle_api_key", &self.gas_oracle_api_key)
             .finish()
@@ -74,26 +59,10 @@ impl Debug for TxManagerConfig {
 pub enum Error {
     #[error("Configuration missing chain_id")]
     MissingChainId,
-
-    #[error("Configuration missing mnemonic")]
-    MissingMnemonic,
-
-    #[error("Could not read mnemonic file at path `{}`: {}", path, source)]
-    MnemonicFileReadError {
-        path: String,
-        source: std::io::Error,
-    },
-
-    #[error("Mnemonic index malformed: {:?}", source)]
-    MnemonicIndexMalformed { source: WalletError },
-
-    #[error("Mnemonic malformed: {0}")]
-    MnemonicMalformed(WalletError),
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-const DEFAULT_MNEMONIC_ACCOUNT_INDEX: u32 = 0;
 const DEFAULT_DEFAULT_CONFIRMATIONS: usize = 7;
 const DEFAULT_HTTP_ENDPOINT: &str = "http://localhost:8545";
 const DEFAULT_DATABASE_PATH: &str = "./default_tx_database";
@@ -117,33 +86,6 @@ impl TxManagerConfig {
         let chain_id = env_cli_config.tx_chain_id.ok_or(Error::MissingChainId)?;
         let chain_is_legacy = env_cli_config.tx_chain_is_legacy.unwrap_or(false);
 
-        let wallet = {
-            let mnemonic: String = if let Some(m) = env_cli_config.tx_mnemonic {
-                m
-            } else {
-                let path = env_cli_config
-                    .tx_mnemonic_file
-                    .ok_or(Error::MissingMnemonic)?;
-
-                let contents = fs::read_to_string(path.clone())
-                    .map_err(|source| Error::MnemonicFileReadError { path, source })?;
-
-                contents.trim().to_string()
-            };
-
-            let index = env_cli_config
-                .tx_mnemonic_account_index
-                .unwrap_or(DEFAULT_MNEMONIC_ACCOUNT_INDEX);
-
-            MnemonicBuilder::<English>::default()
-                .phrase(mnemonic.as_str())
-                .index(index)
-                .map_err(Error::MnemonicMalformed)?
-                .build()
-                .map_err(Error::MnemonicMalformed)?
-                .with_chain_id(chain_id)
-        };
-
         let database_path = env_cli_config
             .tx_database_path
             .unwrap_or(DEFAULT_DATABASE_PATH.to_string());
@@ -155,7 +97,6 @@ impl TxManagerConfig {
         Ok(Self {
             default_confirmations,
             provider_http_endpoint,
-            wallet,
             chain_id,
             chain_is_legacy,
             database_path,